updated VOLTAGE_STATUS to display all values in the CCU web ui
Following values are reported for the VOLTAGE_STATUS of an analog input of a HmIP-MIO16-PCB:1.
Default = "NORMAL"
Max = "OVERFLOW"
Min = "NORMAL"
ValueList = ["NORMAL", "UNKNOWN", "OVERFLOW"]
Even when using these values, the VOLTAGE_STATUS is not displayed in the Web-UI of the CCU, e.g. as a program trigger, as it is for a real device.
Maybe someone can explain.

With the following settings at least all possible values of the enum are displayed, although the value 1 is normally hidden.
Default = 0
Max = 3
Min = 0
ValueList = ["NORMAL", "UNKNOWN", "OVERFLOW", "UNDERFLOW"]